Make get_camera_parameters send the same message as the app
Node telecommunication ignores the second char, so any char can be
used. But if this is fixed in node_telecommunication it is better for the protocol to
be consistent with the app, and the app sends "Va" so only node_telecommunications has to be
changed. Since the protocol is not widely distributed yet and this
does not require updates to the drone it is fine to change this
without a new protocol revision.