Add shell script to generate HTML and Markdown documentation
Script generates documentation based on comments in the .proto files using the protoc-gen-doc command, whish is packaged as a docker container.

See https://github.com/pseudomuto/protoc-gen-doc for details about the tool.

Files are output in an ignored docs folder. The markdown file can be used in the ProtocolDefinition repository Wiki for documentation.